simplify workflow status methods
the workflow status methods reduce the status of multiple jobs of a
workflow to a single status value; the merit of this is debatable.
However, the new Status value object makes this a whole lot easier to
manage. We reduce the status of each of the jobs of a workflow to a
single Status value, and then we use this Status value to call the
various methods