stopping jobs becomes responsibility of simulation
The simulation/workflow object has_many jobs.
To ensure that we qdel all jobs prior to deleting the records of all jobs,
in a before_destroy filter on the simulation we call stop on each job.

If each succeeds without error, we then delete the staged directory.

We removed the before filter on the job objects themselves, which no longer
need to run to stop themselves.